Add unit test to verify that the repository contract has been called with expected parameters
