+++ src/main/kotlin/com/example/tdd_workshop/domain/Report.kt
@@ -1,0 +1,3 @@
+package com.example.tdd_workshop.domain
+
+data class Report(val name: String, val note: Int)
