+++ src/main/kotlin/com/example/tdd_workshop/domain/ReportsRepository.kt
@@ -1,0 +1,5 @@
+package com.example.tdd_workshop.domain
+
+interface ReportsRepository {
+    fun save(report: Report)
+}
